+++ extensions/raxx_logger/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## next
+
+### Changed
+
+- Complete rewrite to use `Raxx.Middleware` behaviour.
+
 ## [0.1.0](https://github.com/CrowdHailer/raxx/tree/0.17.5) - 2019-02-04
 
 ### Added